@@ -2,7 +2,7 @@ package newbilius.com.online_comics_reader.Lists;
 
 import android.app.Activity;
 import android.os.AsyncTask;
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
