@@ -2,7 +2,7 @@ package newbilius.com.online_comics_reader.UI;
 
 import android.content.Context;
 import android.content.DialogInterface;
-import android.support.v7.app.AlertDialog;
+import androidx.appcompat.app.AlertDialog;
 import android.widget.Toast;
 
 public class MessageHelper {
